Modify email adress

+++ ModifyModule/com/example/apimodify/modifyrestapi/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserJPA, Long> {
+public interface UserRepository extends JpaRepository<UserJPA, Integer> {
     Optional<UserJPA> findByUsername(String username);
     Optional<UserJPA> findByEmail(String email);
     Optional<UserJPA> findByUsernameOrEmail(String username, String email);
